+++ src/main/java/org/apiphany/security/Sensitive.java
@@ -1,0 +1,19 @@
+package org.apiphany.security;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Annotation to mark sensitive fields.
+ *
+ * @author Radu Sebastian LAZIN
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ ElementType.FIELD, ElementType.METHOD })
+public @interface Sensitive {
+
+	// empty
+
+}