+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 Added `PRF` with utility method to a apply pseudo-random function in TLS environments.
 Added `KeyExchangeHandler` interface to be implemented in key exchanges.
+Added `Sensitive` annotation to mark sensitive fields.
+Changed `JacksonJsonBuilder` to read sensitive but not write `Sensitive` fields.
 
 ---